# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "API for managing contacts"
  },
  "servers": [
    {
      "url": "https://contacts-app-hw7-swagger.onrender.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Operations about contacts."
    },
    {
      "name": "Auth",
      "description": "Auth operations."
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Retrieve all contacts",
        "operationId": "getAllContacts",
        "description": "Fetch a list of all contacts.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Number of contacts per page."
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Page number."
          },
          {
            "name": "isFavourite",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter contacts by favourite status (true/false)."
          },
          {
            "name": "contactType",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter contacts by contact type."
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Filter contacts by type (Required)."
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "createdAt",
                "updatedAt"
              ]
            },
            "description": "Sort contacts by name, creation date, or update date."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of contacts retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. The request could not be understood or was missing required parameters."
          },
          "404": {
            "description": "Contacts not found."
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "Add a new contact to the list.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the contact"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address of the contact"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone number of the contact"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Mark contact as favourite (true/false)"
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Type of contact (e.g., friend, work, family)"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Profile photo of the contact"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data."
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get a contact by ID",
        "operationId": "getContactById",
        "description": "Fetch detailed information about a contact by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the contact."
          }
        ],
        "responses": {
          "200": {
            "description": "Contact retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Invalid contact ID format."
          },
          "404": {
            "description": "Contact not found."
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update a contact",
        "operationId": "updateContact",
        "description": "Modify details of an existing contact.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the contact"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address of the contact"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone number of the contact"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Mark contact as favourite (true/false)"
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Type of contact (e.g., friend, work, family)"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "New profile photo for the contact"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data."
          },
          "404": {
            "description": "Contact not found."
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete a contact",
        "operationId": "deleteContact",
        "description": "Remove a contact by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact deleted successfully."
          },
          "404": {
            "description": "Contact not found."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Contact": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the contact."
          },
          "email": {
            "type": "string",
            "description": "Email of the contact."
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the contact."
          }
        }
      }
    }
  }
}